+++ lib/src/websocket/models/connection_state.dart
@@ -8,4 +8,13 @@ enum ConnectionState {
   ;
 
   bool get isConnected => this == ConnectionState.connected;
+
+  bool get isConnecting =>
+      this == ConnectionState.connecting ||
+      this == ConnectionState.authenticating;
+
+  bool get isDisconnected =>
+      this == ConnectionState.disconnected ||
+      this == ConnectionState.closing ||
+      this == ConnectionState.closed;
 }