+++ lib/src/websocket/websocket_errors.dart
@@ -44,6 +44,27 @@ class UnexpectedWebsocketException extends Error
   }
 }
 
+class WebsocketClosedError extends Error
+    implements DartactylWebsocketException {
+  WebsocketClosedError([this.message = 'Websocket is closed']);
+
+  @override
+  final String message;
+
+  @override
+  String toString() => 'WebsocketClosedException: $message';
+}
+
+class WebsocketDisconnectedException extends DartactylWebsocketException {
+  WebsocketDisconnectedException([this.message = 'Websocket is disconnected']);
+
+  @override
+  final String message;
+
+  @override
+  String toString() => 'WebsocketDisconnectedException: $message';
+}
+
 class JWTError implements WingsException {
   const JWTError._(String? receivedMessage)
       : receivedMessage = receivedMessage ?? 'Unknown JWT Error';